test(aunit): move coverage-link fixture test to coverage suite
The standalone `it()` in aunit.test.ts bypassed ContractScenario /
ContractOperation and tested no HTTP operation — it was a fixture-
parsing test for extractCoverageMeasurementId. Relocate it to
coverage.test.ts, co-locating it with the existing
extractCoverageMeasurementId unit tests.

Resolves CodeRabbit review comment on PR #175.
